# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TUYUL Trading Bot API
  description: |
    TUYUL is a cryptocurrency trading automation platform for Indodax exchange.

    Features:
    - **Copilot Trading**: Manual trade execution with automated exit strategies
    - **Market Maker Bot**: Spread-capturing automated trading
    - **Pump Hunter Bot**: Momentum-based automated trading
    - **Real-time Market Analysis**: Multi-timeframe pump detection
    - **Paper Trading**: Risk-free simulation mode

    All endpoints require JWT authentication except health checks and auth endpoints.
  version: 1.0.0
  contact:
    name: TUYUL API Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.tuyul.example.com
    description: Production server (example)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid request parameters"

    # User Schemas
...
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: "john_doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          minLength: 8
          example: "SecurePass123!"
...
paths:
...
  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          access_token:
                            type: string
                          expires_in:
                            type: integer
  /api/v1/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/v1/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'

  # User Management Endpoints (Admin)
  /api/v1/users:
    get:
      tags:
        - Users
      summary: List all users (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          users:
                            type: array
                            items:
                              $ref: '#/components/schemas/User'
                          total:
                            type: integer
                          page:
                            type: integer
                          limit:
                            type: integer

    post:
      tags:
        - Users
      summary: Create user (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
  /api/v1/users/profile:
    get:
      tags:
        - Users
      summary: Get user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'

    put:
      tags:
        - Users
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/v1/users/password:
    post:
      tags:
        - Users
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - old_password
                - new_password
              properties:
                old_password:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/v1/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'

    put:
      tags:
        - Users
      summary: Update user (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                role:
                  type: string
                  enum: [admin, user]
                status:
                  type: string
                  enum: [active, inactive, suspended]
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

    delete:
      tags:
        - Users
      summary: Delete user (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/v1/users/{id}/reset-password:
    post:
      tags:
        - Users
      summary: Reset user password (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - new_password
              properties:
                new_password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # API Key Management
  /api/v1/api-keys:
    post:
      tags:
        - API Keys
      summary: Create and validate API key
      description: Stores encrypted Indodax API key and validates it
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/APIKeyRequest'
      responses:
        '201':
          description: API key created and validated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/APIKey'

    get:
      tags:
        - API Keys
      summary: Get API key status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: API key information (encrypted values not returned)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/APIKey'

    delete:
      tags:
        - API Keys
      summary: Delete API key
      security:
        - bearerAuth: []
      responses:
        '200':
          description: API key deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/v1/api-keys/validate:
    post:
      tags:
        - API Keys
      summary: Validate API key with Indodax
      security:
        - bearerAuth: []
      responses:
        '200':
          description: API key is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/v1/api-keys/account-info:
    get:
      tags:
        - API Keys
      summary: Get Indodax account info
      description: Fetches balance and account information from Indodax
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account information
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          balance:
                            type: object
                            additionalProperties:
                              type: string
                          balance_hold:
                            type: object
                            additionalProperties:
                              type: string

  # Market Data Endpoints
  /api/v1/market/summary:
    get:
      tags:
        - Market Data
      summary: Get all market summaries
      description: Returns market data for all trading pairs sorted by pump score
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: min_volume
          in: query
          schema:
            type: number
            default: 0
      responses:
        '200':
          description: Market summaries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          markets:
                            type: array
                            items:
                              $ref: '#/components/schemas/Coin'
                          count:
                            type: integer
                          last_update:
                            type: string
                            format: date-time
  /api/v1/market/pump-scores:
    get:
      tags:
        - Market Data
      summary: Get top pump scores
      description: Returns markets sorted by pump score (highest first)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: min_volume
          in: query
          schema:
            type: number
            default: 1000000
      responses:
        '200':
          description: Top pump scores
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          scores:
                            type: array
                            items:
                              $ref: '#/components/schemas/Coin'
                          count:
                            type: integer
  /api/v1/market/gaps:
    get:
      tags:
        - Market Data
      summary: Get top gaps
      description: Returns markets sorted by bid-ask gap percentage
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: min_volume
          in: query
          schema:
            type: number
            default: 10000000
      responses:
        '200':
          description: Top gaps
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          gaps:
                            type: array
                            items:
                              $ref: '#/components/schemas/Coin'
                          count:
                            type: integer
  /api/v1/market/{pair}:
    get:
      tags:
        - Market Data
      summary: Get pair detail
      description: Returns detailed market data for a specific trading pair
      parameters:
        - name: pair
          in: path
          required: true
          schema:
            type: string
          example: "btcidr"
      responses:
        '200':
          description: Pair details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Coin'
        '404':
          description: Pair not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/market/sync:
    post:
      tags:
        - Market Data
      summary: Sync metadata (Admin only)
      description: Manually triggers metadata refresh from Indodax
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Metadata synced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # Copilot Trading Endpoints
  /api/v1/copilot/trade:
    post:
      tags:
        - Copilot
      summary: Place a copilot trade
      description: |
        Places a buy order with automated sell strategy.
        - Supports both paper trading and live trading
        - Automatically places sell order when buy fills
        - Monitors stop-loss conditions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TradeRequest'
      responses:
        '201':
          description: Trade created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Trade'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/copilot/trades:
    get:
      tags:
        - Copilot
      summary: List copilot trades
      description: Returns all trades for the current user
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, filled, completed, cancelled, stopped]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of trades
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          trades:
                            type: array
                            items:
                              $ref: '#/components/schemas/Trade'
                          count:
                            type: integer
  /api/v1/copilot/trades/{id}:
    get:
      tags:
        - Copilot
      summary: Get trade details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Trade details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Trade'
        '404':
          description: Trade not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Copilot
      summary: Cancel pending trade
      description: Cancels the buy order if trade is still pending
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Trade cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/v1/copilot/trades/{id}/sell:
    post:
      tags:
        - Copilot
      summary: Manual sell
      description: |
        Manually triggers a sell order for a filled trade.
        - Cancels auto-sell order if exists
        - Places market sell order immediately
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Sell order placed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # Bot Management Endpoints
...
  /api/v1/bots/{id}/start:
    post:
      tags:
        - Bots
      summary: Start bot
      description: Starts a stopped bot
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Bot started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/v1/bots/{id}/stop:
    post:
      tags:
        - Bots
      summary: Stop bot
      description: Stops a running bot
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Bot stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...